fix: Bug on view-profile page
*Moved get function from view-profile page to login page
* Please note a known bug, this will not display correctly after adding new recipe unless you logout first
